requirements.txt
chromadb
sentence-transformers
pinecone-client
pymilvus
pyyaml
pytest
qdrant-client 